Refactor Duke class
* Remove Launcher class and put logic in Duke class
This is so that we can launch directly from the Duke class and not need
a separate class just to launch the GUI.
* Move Main class to gui package
This is to make the application more organised.
